Refactor ISVStepper context and update imports across components. Add button type to ISVApplyActions for better accessibility. Adjust resolver type in ISVConfiguration for improved type safety.
