+++ src/react/ISV/components/ISVSteps.tsx
@@ -1,8 +1,6 @@
 import { spacing, Stepper } from '@scality/core-ui';
-import { createContext, useContext, useMemo } from 'react';
+import { useMemo } from 'react';
 import { ISVConfiguration } from './ISVConfiguration';
-
-import { ISVPlatformConfig } from '../types';
 
 import { useTheme } from 'styled-components';
 import { Box } from '@scality/core-ui/dist/next';
@@ -11,6 +9,7 @@ import ISVApplyActions from './ISVApplyActions';
 
 import { useSearchParams } from 'react-router';
 import { isvModules } from '../modules';
+import { ISVStepperContext } from './ISVStepperContext';
 
 
 export enum ISVStepsIndexes {
@@ -18,22 +17,6 @@ export enum ISVStepsIndexes {
   ApplyActions,
   Summary,
 }
-
-export type ISVStepperContextType = {
-  platform: ISVPlatformConfig;
-};
-
-export const ISVStepperContext = createContext<
-  ISVStepperContextType | undefined
->(undefined);
-
-export const useISVStepper = () => {
-  const context = useContext(ISVStepperContext);
-  if (!context) {
-    throw new Error('useISVStepper must be used within ISVStepperProvider');
-  }
-  return context;
-};
 
 export const ISV_STEPS = [
   {
